+++ test/pytests/test_smart_completion_public_schema_only.py
@@ -80,7 +80,13 @@ def complete_event():
 def test_use_database_completion(completer, complete_event):
     text = "USE "
     position = len(text)
-    special.register_special_command(..., 'use', '\\u [database]', 'Change to a new database.', aliases=['\\u'])
+    special.register_special_command(
+        ...,
+        'use',
+        '\\u [database]',
+        'Change to a new database.',
+        aliases=[special.SpecialCommandAlias('\\u', case_sensitive=False)],
+    )
     result = completer.get_completions(Document(text=text, cursor_position=position), complete_event)
     assert list(result) == [
         Completion(text="test", start_position=0),
@@ -652,7 +658,13 @@ def dummy_list_path(dir_name):
 )
 def test_file_name_completion(completer, complete_event, text, expected):
     position = len(text)
-    special.register_special_command(..., 'source', '\\. <filename>', 'Execute commands from file.', aliases=['\\.'])
+    special.register_special_command(
+        ...,
+        'source',
+        '\\. <filename>',
+        'Execute commands from file.',
+        aliases=[special.SpecialCommandAlias('\\.', case_sensitive=False)],
+    )
     result = list(completer.get_completions(Document(text=text, cursor_position=position), complete_event))
     expected = [Completion(txt, pos) for txt, pos in expected]
     assert result == expected
@@ -689,7 +701,13 @@ def test_source_eager_completion(completer, complete_event, tmp_path, monkeypatc
     script_filename = 'do_these_statements.sql'
     f = open(script_filename, 'w')
     f.close()
-    special.register_special_command(..., 'source', '\\. <filename>', 'Execute commands from file.', aliases=['\\.'])
+    special.register_special_command(
+        ...,
+        'source',
+        '\\. <filename>',
+        'Execute commands from file.',
+        aliases=[special.SpecialCommandAlias('\\.', case_sensitive=False)],
+    )
     result = list(completer.get_completions(Document(text=text, cursor_position=position), complete_event))
     success = True
     error = 'unknown'
@@ -715,7 +733,13 @@ def test_source_leading_dot_suggestions_completion(completer, complete_event, tm
     script_filename = 'do_these_statements.sql'
     f = open(script_filename, 'w')
     f.close()
-    special.register_special_command(..., 'source', '\\. <filename>', 'Execute commands from file.', aliases=['\\.'])
+    special.register_special_command(
+        ...,
+        'source',
+        '\\. <filename>',
+        'Execute commands from file.',
+        aliases=[special.SpecialCommandAlias('\\.', case_sensitive=False)],
+    )
     result = list(completer.get_completions(Document(text=text, cursor_position=position), complete_event))
     success = True
     error = 'unknown'